requirements.txt
numpy
pandas
scikit-learn
pymysql
mysql-connector-python
python-dotenv
seaborn
matplotlib
xgboost
catboost

-e .